src/server/builder/requirements.txt
fastmcp>=0.9.0